requirements.txt
SQLAlchemy==1.4.54
alembic==1.13.3
psycopg2-binary==2.9.11
requests==2.31.0
python-telegram-bot==13.15
prometheus-client==0.21.1
pytest==8.3.4
pytest-mock==3.14.0
pyhamcrest==2.1.0
